chore(nlu): fully removed spellchecking from prediction
