+++ src/bp/nlu-core/typings.ts
@@ -150,5 +150,3 @@ type SlotDefinition = Readonly<{
   name: string
   entities: string[]
 }>
-
-export type ValueOf<T> = T[keyof T]
